249: Use flutter theme
