--- lib/app/constants/theme_colors.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ThemeColors {
-  static const Color middleGrey = Color(0xFF9CABAF);
-  static const Color stateHoverGreen = Color(0xFF005437);
-  static const Color white = Color(0xFFFFFFFF);
-}
