--- lib/app/constants/theme_text_styles.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-
-class ThemeTextStyles {
-  static TextStyle smallText = GoogleFonts.ptSans(
-    textStyle: TextStyle(
-      fontSize: 10,
-      fontWeight: FontWeight.w400,
-      height: 1.3,
-      letterSpacing: 0.01,
-    ),
-  );
-}
